dynamic url for recent and featured card
