+++ src/components/ProjectCard.scss
@@ -8,6 +8,7 @@
 
 .project-card {
   width: 320px;
+  margin-top: $large;
   &__header {
     background: $lightPurple;
     padding: $small $medium;
